Array for word + Random picker
I now have a working code that picks a random word for my program.
Progress has been made, my next goal is to figure out how I would use a
command that figures out and analyses the word to see how many letters
are in it as well as which letters it includes.