Another Revision
I have never had so much trouble with a code. The problem that I keep
on having is the increment of drawHangman is increasing by more than 1
overtime because I can’t get the key to only be held down for a
milasecond. Anyways lots of work done here, if you care to take a peak.